Moved OpenGisMapper to netex package
