+++ src/main/java/org/opentripplanner/netex/loader/mapping/FlexStopLocationMapper.java
@@ -1,6 +1,5 @@
 package org.opentripplanner.netex.loader.mapping;
 
-import org.opentripplanner.common.geometry.OpenGisMapper;
 import org.opentripplanner.model.FlexStopLocation;
 import org.rutebanken.netex.model.FlexibleArea;
 import org.rutebanken.netex.model.FlexibleStopPlace;
